Add API to unregister nodes from keyboard visibility changes
