+++ modules/keyboard/src/main/java/com/gluonhq/attach/keyboard/impl/IOSKeyboardService.java
@@ -30,10 +30,8 @@ package com.gluonhq.attach.keyboard.impl;
 import com.gluonhq.attach.lifecycle.LifecycleEvent;
 import com.gluonhq.attach.lifecycle.LifecycleService;
 import javafx.application.Platform;
-import javafx.beans.property.ReadOnlyFloatProperty;
 import javafx.beans.property.ReadOnlyStringProperty;
 import javafx.scene.Node;
-import javafx.scene.Parent;
 import javafx.scene.control.TextInputControl;
 
 public class IOSKeyboardService extends BaseKeyboardService {
@@ -54,20 +52,6 @@ public class IOSKeyboardService extends BaseKeyboardService {
         startObserver();
     }
 
-    @Override
-    public void keepVisibilityForNode(Node node) {
-        keepVisibilityForNode(node, null);
-    }
-
-    @Override
-    public void keepVisibilityForNode(Node node, Parent parent) {
-        VISIBLE_HEIGHT.addListener((obs, ov, nv) -> adjustPosition(node, parent, nv.doubleValue()));
-    }
-
-    @Override
-    public ReadOnlyFloatProperty visibleHeightProperty() {
-        return VISIBLE_HEIGHT.getReadOnlyProperty();
-    }
 
     @Override
     protected void applyKeyboardType(int nativeValue) {
